Add Numa balancing and logging
By default, numa balancing will be off.  While this has been well
tested for a specific configuration, it has not been tested across
a large number of configurations, and could cause serious
performance impacts.  Also, it may be necessary in certain configs
to exclude certain numa groups from RXE allocations.

Numa logging uses hw_counters and is limited to 8 "bins".  This is
purely to limit the number of required counters and may be enhanced
in the future if we find many use cases with more than 8 numa groups.